local_requirements.txt
-r requirements.txt
django-debug-toolbar==1.5
